From: Tai, Yu-Tsung Sent: Thursday, November 05, 2015 5:33 To: Hanson, Andrew J. <hansona@indiana.edu> Cc: Amr A. Sabry (sabry@indiana.edu) <sabry@indiana.edu>; 'Gerardo Ortiz' <ortizg@indiana.edu> Subject: RE: product, unentangled, entangled, maximally entangled
Actually, I have wrote a program to test whether a state is a product
state or not as attachment... We may check whether a state is a product
state (with a particular type) for a given p.

The state we are interested in is \ket{01} + \ket{10} + (1 - i) \ket{11}
for p = 3, and I am still thinking the meaning of it...

Best regards,
Tai, Yu-Tsung (戴淯琮)
PhD student in Mathematics & Computer Science Indiana University,
Bloomington

-----Original Message-----
From: Hanson, Andrew J. 
Sent: Monday, November 02, 2015 14:45
To: Tai, Yu-Tsung <yuttai@indiana.edu>
Subject: RE: product, unentangled, entangled, maximally entangled

Did you determine whether or not the Purity=1 entangled state could be
written also as a product state?  We mentioned that, then dropped it and
moved on to the n=p question before I understood the answer.
    Andy